Healthcare_booking/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Flask-Migrate==3.1.0
python-dotenv==0.19.2
psycopg2-binary==2.9.3
pyjwt==2.6.0
